refactor(App.vue): extract necessary settings watchers to dedicated file
